rename: 코로케이션 적용
