+++ src/pages/SignUp/hooks/useSignUpQuery.ts
@@ -1,7 +1,7 @@
-import { useQuery } from '@tanstack/react-query';
+import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { getSingUp } from '../../api';
-import { convertCheckboxQuery, convertSkillQuery } from '../../service/convertSignUpQuery';
+import { getSingUp } from '../../../api';
+import { convertCheckboxQuery, convertSkillQuery } from '../../../service/convertSignUpQuery';
 
 const SKILL_DEPTH_THREE_LIST = [
   { id: 1, name: '상' },
@@ -29,13 +29,13 @@ const REGION_LIST = [
 ];
 
 const useSignUpQuery = () => {
-  const { data } = useQuery({
+  const { data } = useSuspenseQuery({
     queryKey: ['GetSignUp'],
     queryFn: getSingUp,
   });
 
-  const { mainSkillQuery, detailSkillQuery } = convertSkillQuery(data?.mainSkillResponses);
-  const checkboxQuery = convertCheckboxQuery(data?.purposeResponses);
+  const { mainSkillQuery, detailSkillQuery } = convertSkillQuery(data.mainSkillResponses);
+  const checkboxQuery = convertCheckboxQuery(data.purposeResponses);
 
   return {
     mainSkillQuery,
